Create script to build xoop for different shaped screens
- refactor to define shape and rules in map_* scripts
